feat(autograph): Complete Phase 1 - Native UI Foundation
Implemented full egui-based visual flow editor for Autograph with:

**Core Features:**
- Native desktop app (egui + eframe)
- Visual canvas with grid, pan, and zoom
- Node palette with 6 node types (start, print, http_request, json_parse, tensor_create, tensor_op)
- Drag-and-drop node creation
- Visual edge drawing with bezier curves (Ctrl+Click to connect)
- Properties panel with JSON config editor
- Save/Load flows as JSON
- Compile button (generates .hlxa files)
- Run button (executes flows with HLX runtime)
- Delete nodes with Delete key
- Node selection and highlighting

**Architecture:**
- src/ui.rs - Main app structure
- src/ui/canvas.rs - Canvas rendering and interaction
- src/ui/palette.rs - Node palette panel
- src/ui/properties.rs - Properties editor panel
- CLI: autograph ui (default) | autograph server --port 3000

**Integration:**
- Reuses existing Flow → HLX compiler from flow.rs
- Direct integration with hlx_compiler and hlx_runtime
- Full deterministic execution via HLX
- JSON input/output for flows

Phase 1 MVP complete - ready for Phase 2 (Core Node Library expansion).

+++ autograph/src/flow.rs
@@ -22,10 +22,10 @@ pub struct Edge {
     pub target_handle: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Flow {
